refactor user model
